Deduplicate the gallery pick+crop flow into a shared helper
Follow-up from a thermo-nuclear code quality review of the previous
commit, which flagged two duplication issues:

- pickImageFromGallery was copy-pasted verbatim between utils/file.ts
  and utils/file.web.ts even though it has no platform-specific
  behavior (unlike openCropperAsync, which genuinely differs).
  Moves it into a new utils/galleryImagePicker.ts with no .web.ts
  counterpart.

- CoachModal, CreateMealModal, CreateCustomFoodModal, and
  CreateExerciseModal each repeated the same pick -> null-check ->
  crop -> null-check block. Collapses this into a single
  pickAndCropImageFromGallery helper (also in
  utils/galleryImagePicker.ts) that every one of the four modals now
  calls directly. This also closes a real gap the duplication had
  produced: CreateExerciseModal's handler didn't wrap the pick/crop
  calls in a try/catch like the other three did, so a picker or
  crop-tool failure there was an unhandled rejection with no
  user-facing error; it now shares the same error handling.

useCameraCaptureFlow.ts keeps calling pickImageFromGallery and
openCropperAsync separately, since it needs distinct phase-timing
logs and error messages per step.
